backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.27.0

# Machine Learning
pandas>=2.0,<3.0
xgboost>=2.0,<3.0
scikit-learn>=1.4,<2.0
numpy>=1.26,<2.0
joblib>=1.3,<2.0
